+++ statistics-node/src/main/java/com/dwarfeng/statistics/node/configuration/ServiceExceptionMapperConfiguration.java
@@ -49,11 +49,16 @@ public class ServiceExceptionMapperConfiguration {
         destination.put(FilterException.class, ServiceExceptionCodes.FILTER_FAILED);
         destination.put(FilterMakeException.class, ServiceExceptionCodes.FILTER_MAKE_FAILED);
         destination.put(FilterExecutionException.class, ServiceExceptionCodes.FILTER_EXECUTION_FAILED);
+        destination.put(UnsupportedFilterTypeException.class, ServiceExceptionCodes.FILTER_TYPE_UNSUPPORTED);
         destination.put(InvalidVariableValueTypeException.class, ServiceExceptionCodes.INVALID_VARIABLE_VALUE_TYPE);
         destination.put(VariableNotExistsException.class, ServiceExceptionCodes.VARIABLE_NOT_EXISTS);
         destination.put(VariableValueTypeMismatchException.class, ServiceExceptionCodes.VARIABLE_VALUE_TYPE_MISMATCH);
         destination.put(ProviderDataExceededException.class, ServiceExceptionCodes.PROVIDER_DATA_EXCEEDED);
         destination.put(QueryException.class, ServiceExceptionCodes.QUERY_FAILED);
+        destination.put(MapperException.class, ServiceExceptionCodes.MAPPER_FAILED);
+        destination.put(MapperMakeException.class, ServiceExceptionCodes.MAPPER_MAKE_FAILED);
+        destination.put(MapperExecutionException.class, ServiceExceptionCodes.MAPPER_EXECUTION_FAILED);
+        destination.put(UnsupportedMapperTypeException.class, ServiceExceptionCodes.MAPPER_TYPE_UNSUPPORTED);
         return new MapServiceExceptionMapper(destination, com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.UNDEFINED);
     }
 }